+++ CHANGELOG.md
@@ -13,14 +13,15 @@ All user visible changes to `cucumber` crate will be documented in this file. Th
 
 ### BC Breaks
 
-- Added `event::ScenarioFinished` argument to after hook handler. ([#246], [#245])
+- Added `event::ScenarioFinished` as [`Cucumber::after`][0170-1] hook's argument, explaining why the `Scenario` has finished. ([#246], [#245])
 
 ### Fixed
 
-- Catch panics in user code, when they happen before first `Future::poll()` (on `Future` construction). ([#246])
+- Uncaught panics of user code, when they happen before first poll of the returned `Future`s. ([#246])
 
-[#245]: /../../issues/245
+[#245]: /../../discussions/245
 [#246]: /../../pull/246
+[0170-1]: https://docs.rs/cucumber/0.17.0/cucumber/struct.Cucumber.html#method.after